+++ PCLParaphernalia/ToolSymbolSetGenPCL.cs
@@ -107,7 +107,7 @@ namespace PCLParaphernalia
                     //--------------------------------------------------------//
 
                     PCLWriter.symSetDownloadCode (_binWriter,
-                                                  (ushort)symSetNo);
+                                                  symSetNo);
 
                     //--------------------------------------------------------//
                     //                                                        //
